add regexp verifications

+++ backend/controllers/sauces.js
@@ -18,7 +18,6 @@ exports.getOneSauce = (req, res, next) => {
 exports.createSauce = (req, res, next) => {
     const sauceObject = JSON.parse(req.body.sauce);
     delete sauceObject._id;
-    if(!/[^<>$#+]/.test(Object.values(sauceObject).join(''))){
     const sauce = new Sauce({
         ...sauceObject,
         imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`,
@@ -29,28 +28,29 @@ exports.createSauce = (req, res, next) => {
     });
     sauce.save()
         .then(() => { res.status(201).json({ message: 'Sauce enregistrée avec succés !' }); })
-        .catch((error) => { res.status(400).json({ error: error.message }); });}
-        else{
-            res.status(400).json({message:"Veuillez vérifier que vos entrées ne contiennent pas de caractères spéciaux."});
-        }
-};
+        .catch((error) => { res.status(400).json({ error: error.message }); });
+
+}
+
 
 //put/:id
 exports.modifySauce = (req, res, next) => {
-    Sauce.findById(req.params.id)
-        .then(sauce => {
-            const filename = sauce.imageUrl.split('/images/')[1];
-            fs.unlink(`images/${filename}`, () => { console.log("image supprimée") });
-        });
+    if (req.file) {
+        Sauce.findById(req.params.id)
+            .then(sauce => {
+                const filename = sauce.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => { console.log("image supprimée") });
+            });
+    }
     const sauceObject = req.file ?
         {
             ...JSON.parse(req.body.sauce),
             imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
         } : { ...req.body };
 
-    Sauce.findByIdAndUpdate(req.params.id, { ...sauceObject, _id: req.params.id })
+    Sauce.findByIdAndUpdate(req.params.id, { ...sauceObject })
         .then(() => res.status(200).json({ message: 'Sauce modifiée !' }))
-        .catch(error => res.status(400).json({ error }));
+        .catch(error => res.status(400).json({ error: error.message }));
 };
 
 //delete/:id
@@ -103,7 +103,7 @@ exports.addLikeToSauce = (req, res, next) => {
         }
 
         //...sauce do not work ?
-        Sauce.findByIdAndUpdate(req.params.id, { ...sauce})
+        Sauce.findByIdAndUpdate(req.params.id, { ...sauce })
             .then(() => res.status(201).json({ message: 'Votre vote a bien été enregistré !' }))
             .catch(error => res.status(400).json({ error }));
     })
